Extract tooltip
